llm-data-quality/requirements_cpu.txt
pydantic==1.9.1
transformers==4.25.1
evaluate==0.4.0
tqdm==4.64.1
pyyaml==6.0
faiss-cpu
datasets==2.8.0
git+https://github.com/flairNLP/flair
gensim==4.3.0